Update PackageReference to latest version
Hi, many thanks for the tutorial. I followed it along and noticed that there were errors when targeting the specified beta version of Veldrid. The errors where that the type DeviceBuffer did not exist, likewise the SubmitCommands method. 'GraphicsBackend.Metal' was also not recognised.
Once the package updated everything worked and it was nice to see some graphics output as I've been reading a few different ways of rendering graphics with .Net Core and some of them haven't worked as yet (or at least haven't had good tutorials).